+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This changelog covers what's changed in the MPDX APIs.
 ## 4 October 2017
 - Contact Status filter now returns contacts with no status when 'active' is in the list of parameter.
 
+## 3 October 2017
+- convert pledge status booleans to enum
+
 ## 29 September 2017
 - Added :pledged_to_appeal filter to appeal/:id/appeal_contacts endpoint.